+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master
 
+* `type_predicates` – Add new plugin with convenient `UploadedFile` predicate methods based on MIME type (@janko)
+
 * `core` – Raise `Shrine::NotCached` in `Attacher#attach_cached` when non-cached uploaded file is assigned (@janko)
 
 * `allow_reassign` – Add new plugin for ignoring exception when assigning stored file if it's the current file (@janko)